apps/parking_offstreet/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.2
PyJWT==2.9.0
python-dotenv==1.0.1
